fix(lens): one home for the host-scoped parent-snapshot memo
Extract createHostScopedParentSnapshot (host-tag + seed + equals gate) into a neutral parentSnapshot.ts holding the relocated sameParentSnapshot comparator; reconcile and adopt each pass their own map projection instead of copy-pasting the memo wrapper.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-3, raised by lowy). Not pushed or merged.
